# backend/src/TechbodiaNotes.Api/DTOs/Notes/NoteResponse.cs
namespace TechbodiaNotes.Api.DTOs.Notes;

public class NoteResponse
{
    public Guid Id { get; set; }
    public Guid UserId { get; set; }
    public string Title { get; set; } = string.Empty;
    public string Content { get; set; } = string.Empty;
    public DateTime CreatedAt { get; set; }
    public DateTime UpdatedAt { get; set; }
}