# specs/001-notes-crud/contracts/api.yaml
openapi: 3.0.3
info:
  title: Techbodia Notes API
  description: RESTful API for Notes CRUD operations
  version: 1.0.0
  contact:
    name: Techbodia
    email: support@techbodia.com

servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.techbodia.com/api
    description: Production server

tags:
  - name: Notes
    description: Notes CRUD operations

paths:
  /notes:
    get:
      tags:
        - Notes
      summary: Get all notes for the current user
      description: Returns a list of notes owned by the authenticated user with optional filtering and sorting
      operationId: getNotes
      parameters:
        - name: search
          in: query
          description: Search text to filter notes by title or content
          required: false
          schema:
            type: string
            maxLength: 100
        - name: sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            enum: [createdAt, updatedAt, title]
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort direction
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          description: Page number (1-based)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Successfully retrieved notes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotesListResponse'
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Notes
      summary: Create a new note
      description: Creates a new note for the authenticated user
      operationId: createNote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteRequest'
      responses:
        '201':
          description: Note created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteResponse'
        '400':
          description: Invalid request - validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /notes/{id}:
    get:
      tags:
        - Notes
      summary: Get a specific note by ID
      description: Returns a single note if it belongs to the authenticated user
      operationId: getNoteById
      parameters:
        - $ref: '#/components/parameters/NoteId'
      responses:
        '200':
          description: Successfully retrieved note
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteResponse'
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - note belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Notes
      summary: Update an existing note
      description: Updates a note if it belongs to the authenticated user
      operationId: updateNote
      parameters:
        - $ref: '#/components/parameters/NoteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNoteRequest'
      responses:
        '200':
          description: Note updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoteResponse'
        '400':
          description: Invalid request - validation errors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - note belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Notes
      summary: Delete a note
      description: Permanently deletes a note if it belongs to the authenticated user
      operationId: deleteNote
      parameters:
        - $ref: '#/components/parameters/NoteId'
      responses:
        '204':
          description: Note deleted successfully
        '401':
          description: Unauthorized - authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - note belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  parameters:
    NoteId:
      name: id
      in: path
      description: Unique identifier of the note
      required: true
      schema:
        type: integer
        minimum: 1

  schemas:
    NoteResponse:
      type: object
      required:
        - id
        - title
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          description: Unique identifier
          example: 1
        title:
          type: string
          description: Note title
          maxLength: 200
          example: "Meeting Notes"
        content:
          type: string
          nullable: true
          description: Note content
          maxLength: 50000
          example: "Discuss project timeline and deliverables."
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (UTC)
          example: "2026-01-17T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp (UTC)
          example: "2026-01-17T14:45:00Z"

    NotesListResponse:
      type: object
      required:
        - notes
        - totalCount
      properties:
        notes:
          type: array
          items:
            $ref: '#/components/schemas/NoteResponse'
        totalCount:
          type: integer
          description: Total number of notes (for pagination)
          example: 42

    CreateNoteRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Note title (required)
          minLength: 1
          maxLength: 200
          example: "New Note"
        content:
          type: string
          nullable: true
          description: Note content (optional)
          maxLength: 50000
          example: "Note content goes here..."

    UpdateNoteRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Note title (required)
          minLength: 1
          maxLength: 200
          example: "Updated Title"
        content:
          type: string
          nullable: true
          description: Note content (optional)
          maxLength: 50000
          example: "Updated content..."

    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
          example: "Note not found"
        code:
          type: string
          description: Error code for programmatic handling
          example: "NOT_FOUND"

    ValidationErrorResponse:
      type: object
      required:
        - message
        - errors
      properties:
        message:
          type: string
          description: General error message
          example: "Validation failed"
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Field-specific validation errors
          example:
            title:
              - "Title is required"
              - "Title must be 200 characters or less"

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication (optional for junior scope)

# Security is optional for junior position scope
# Uncomment to enable authentication requirement
# security:
#   - BearerAuth: []